Rename dbtx package so it can contain database tests
