+++ internal/dbx/querier.go
@@ -1,5 +1,5 @@
-// dbtx contains the [Querier] and [Queries] types, which are maintained manually. The contents of the [db] package are generated by sqlc and regularly overwritten, making it unsuitable for custom code.
-package dbtx
+// dbx extends the [db] package with application-specific database code. The contents of the [db] package are generated by sqlc and regularly overwritten, making it unsuitable for custom code.
+package dbx
 
 import (
 	"github.com/jackc/pgx/v5"
