+++ main.go
@@ -22,7 +22,7 @@ import (
 	"github.com/cloud-gov/billing/internal/api"
 	"github.com/cloud-gov/billing/internal/config"
 	"github.com/cloud-gov/billing/internal/db"
-	"github.com/cloud-gov/billing/internal/dbtx"
+	"github.com/cloud-gov/billing/internal/dbx"
 	"github.com/cloud-gov/billing/internal/jobs"
 	"github.com/cloud-gov/billing/internal/migrate"
 	"github.com/cloud-gov/billing/internal/server"
@@ -86,7 +86,7 @@ func run(ctx context.Context, out io.Writer) error {
 		return fmtErr(ErrDBMigration, err)
 	}
 
-	q := dbtx.NewQuerier(db.New(conn))
+	q := dbx.NewQuerier(db.New(conn))
 
 	logger.Debug("run: initializing meters")
 	meters := []reader.Meter{
